% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions_targetHub.R
\name{miRNA_target_interactions}
\alias{miRNA_target_interactions}
\title{targetHub is a database of miRNA-mRNA interactions. The interaction data is
obtained various external data sources and in some cases computed in-house
by algorithms implemented for miRNA target prediction.}
\usage{
<<<<<<< HEAD
miRNA_target_interactions(query_alt, gene, search_option, data_source)
=======
miRNA_target_interactions(query_alt, search_option, gene, data_source)
>>>>>>> 2f30562a8598120de1d5de5805ad4de22e11edaf
}
\arguments{
\item{query_alt}{gene: Query by gene
stem-loop miRNA: Query by stem-loop miRNA
mature miRNA: Query by mature miRNA}

<<<<<<< HEAD
\item{gene}{geneID character \emph{hsa-miR-212-3p} or numeric \emph{672}}

\item{search_option}{evidence count: Evidence Count
specific method: Specific Methods}

=======
\item{search_option}{evidence count: Evidence Count
specific method: Specific Methods}

\item{gene}{geneID character \emph{hsa-miR-212-3p} or numeric \emph{672}}

>>>>>>> 2f30562a8598120de1d5de5805ad4de22e11edaf
\item{data_source}{character \emph{miranda+mirtarbase+pictar4+targetscan} or numeric \emph{3}}
}
\value{
targetHub object.
}
\description{
targetHub is a database of miRNA-mRNA interactions. The interaction data is
obtained various external data sources and in some cases computed in-house
by algorithms implemented for miRNA target prediction.
}
\examples{
tmp = miRNA_target_interactions("gene", "evidence count", 672,  3)
tmp = miRNA_target_interactions("stem-loop miRNA", "evidence count", "hsa-mir-212", 4)
tmp = miRNA_target_interactions("mature miRNA", "evidence count", "hsa-miR-212-3p", 4)
tmp = miRNA_target_interactions("gene", "Specific Method", 672,  "miranda+pictar4+targetscan")
# supported by exactly three data sources/methods
x = tmp$extract()
# supported by atleast three data sources/methods
x = tmp$atleast()
}

