\name{miRNAtargetpackage-package}
\alias{miRNAtargetpackage-package}
\alias{miRNAtargetpackage}
\docType{package}
\title{
\packageTitle{miRNAtargetpackage}
}
\description{
\packageDescription{miRNAtargetpackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{miRNAtargetpackage}
\packageIndices{miRNAtargetpackage}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{miRNAtargetpackage}

Maintainer: \packageMaintainer{miRNAtargetpackage}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
