% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions_targetHub.R
\name{print.targetHub}
\alias{print.targetHub}
\title{Function that prints the interactions.}
\usage{
\method{print}{targetHub}(x, ...)
}
\arguments{
\item{x}{targetHub object}

\item{...}{other arguments}
}
\description{
Function that prints the interactions.
}

